third_party/blink/web_tests/fast/files/file-reader-file-url.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.setAllowFileAccessFromFileURLs(false);
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<iframe src="resources/file-reader-real.html"></iframe>
<div id="results"></div>
Test that FileReader works under the file URL with file path separation enforced.
</body>
</html>
