third_party/blink/web_tests/fast/files/file-reader-file-url-expected.txt

CONSOLE MESSAGE: Received data: 0123456789
Test that FileReader works under the file URL with file path separation enforced.
